client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles",
    "aiohttp",
    "loguru>=0.7.3",
    "nuitka",
    "pydantic",
    "pywebview",
]
